+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add energy conversion from LNRG and SNRG to VNRG
+- Add `pallet-vitreus-dex` on testnet
+- Add `pallet-launchpad` on testnet
 
 ### Changed